+++ src/main/kotlin/blackjack/domain/Pattern.kt
@@ -1,0 +1,8 @@
+package blackjack.domain
+
+enum class Pattern(val pattern: String) {
+    HEART("하트"),
+    DIAMOND("다이아"),
+    SPADE("스페이드"),
+    CLOVER("클로버")
+}
